Set physics tickrate to 30
Recent change in schedule in avian reduced the performance if the
game underperforms.
On my machine in debug the game underperforms and the slowdown
is noticable. Changing the tickrate to 30 fixes it.
We use physics only for raycasting and collision detection anyway,
there is no reason for to run physics at 60 hz,
better to save the performance for something else.